upcasting array for length plots to avoid int overflow in weighted hist
should fix issue raised in 
https://github.com/wdecoster/NanoPlot/issues/111, in the case of all 
reads shorter than 64kb

+++ nanoplot/NanoPlot.py
@@ -137,7 +137,7 @@ def make_plots(datadf, settings):
         n50 = None
     plots.extend(
         nanoplotter.length_plots(
-            array=datadf[datadf["length_filter"]]["lengths"],
+            array=datadf[datadf["length_filter"]]["lengths"].astype('uint64'),
             name="Read length",
             path=settings["path"],
             n50=n50,
